% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_multiply.R
\name{cross_multiply}
\alias{cross_multiply}
\title{This function is used to multiply across columns of dataframes.}
\usage{
cross_multiply(multiplied, multiplier, variable)
}
\arguments{
\item{multiplied}{The dataframe that to be multiplied.}

\item{multiplier}{The dataframe that you want to multiply. It must be a dataframe with only one row.}

\item{variable}{The variable that you want to multiply.}
}
\value{
The dataframe that has been multiplied.
}
\description{
This function is used to multiply across columns of dataframes.
}
\examples{
cross_multiply(data.frame(a = c(1,2,3),b = c(4,5,6)),data.frame(a = 1,b = 2),a)
}
