% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonaqi.R
\name{fibonaqi}
\alias{fibonaqi}
\title{This function is used to calculate the sum of a vector in a Fibonacci sequence.}
\usage{
fibonaqi(vector)
}
\arguments{
\item{vector}{The vector that you want to calculate.}
}
\value{
The sum of the vector in a Fibonacci sequence.
}
\description{
This function is used to calculate the sum of a vector in a Fibonacci sequence.
}
\examples{
fibonaqi(c(1,1,1,1,1,1,1,1,1,1))
}
