% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name.R
\name{get_name}
\alias{get_name}
\title{Get the file name from the full path}
\usage{
get_name(x, i, suf)
}
\arguments{
\item{x}{The full path}

\item{i}{The index of the file name}

\item{suf}{The suffix of the file name}
}
\value{
The file name
}
\description{
Get the file name from the full path
}
\examples{
get_name('C:/Users/username/Desktop/abc.txt',5,'.txt')
}
