% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygontotaster.R
\name{polygontotaster}
\alias{polygontotaster}
\title{This function is used to convert polygon .shp data to raster data.}
\usage{
polygontotaster(x, res, field, fun)
}
\arguments{
\item{x}{The polygon .shp data}

\item{res}{The resolution of the raster data}

\item{field}{The field of the polygon data}

\item{fun}{The function used in the conversion process, see the help of fasterize::fasterize for details.}
}
\value{
The raster data
}
\description{
This function is used to convert polygon .shp data to raster data.
}
\examples{
polygontotaster('abc.shp',100,'pop','mean')
}
